+++ tasks/ai/article/examples/first/FirstProblemLossTask.py
@@ -3,7 +3,7 @@ from equations.ai.article.examples.first.FirstProblemLoss import FirstProblemLos
 
 
 class FirstProblemLossTask(TaskData):
-    def __init__(self,weight: int = 10):
+    def __init__(self,weight: float = 10):
         super().__init__(SpaceRanges(10, Range(0, 1)),"1 problem loss",weight)
 
     def get_equation(self):
